requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.1.0
gunicorn==22.0.0
lxml==5.3.0
pymongo==4.15.5
flask-cors==4.0.0
google-genai==0.2.2